Code refactor based on PR review
