+++ jbpm-benchmarks/jbpm-performance-tests-jmh/pom.xml
@@ -33,13 +33,13 @@ THE POSSIBILITY OF SUCH DAMAGE.
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-
   <parent>
     <artifactId>jbpm-benchmarks</artifactId>
     <groupId>org.jbpm</groupId>
-    <version>7.40.0-SNAPSHOT</version>
+    <version>7.41.0-SNAPSHOT</version>
   </parent>
-
+  <artifactId>jbpm-performance-tests-jmh</artifactId>
+  <name>JMH benchmark sample: Java</name>
 
   <dependencyManagement>
     <dependencies>
@@ -80,18 +80,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
       <artifactId>kie-internal</artifactId>
     </dependency>
     <dependency>
-      <groupId>io.dropwizard.metrics</groupId>
-      <artifactId>metrics-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.dropwizard.metrics</groupId>
-      <artifactId>metrics-jvm</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jbpm</groupId>
-      <artifactId>jbpm-persistence-jpa</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.jbpm</groupId>
       <artifactId>jbpm-human-task-core</artifactId>
     </dependency>
@@ -114,10 +102,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
     <dependency>
       <groupId>org.jbpm</groupId>
       <artifactId>jbpm-flow</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.drools</groupId>
-      <artifactId>drools-workbench-models-guided-dtable</artifactId>
     </dependency>
     <dependency>
       <groupId>com.h2database</groupId>
@@ -162,12 +146,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
     </dependency>
   </dependencies>
 
-  <artifactId>jbpm-performance-tests-jmh</artifactId>
-  <version>1.0</version>
-  <packaging>jar</packaging>
-
-  <name>JMH benchmark sample: Java</name>
-
   <build>
     <resources>
       <resource>
@@ -183,7 +161,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-shade-plugin</artifactId>
-        <version>3.2.0</version>
         <executions>
           <execution>
             <phase>package</phase>
@@ -191,14 +168,12 @@ THE POSSIBILITY OF SUCH DAMAGE.
               <goal>shade</goal>
             </goals>
             <configuration>
-              <finalName>benchmarks</finalName>
+              <finalName>jbpm-benchmarks</finalName>
               <transformers>
-                <transformer
-                    implementation="org.apache.maven.plugins.shade.resource.AppendingTransformer">
+                <transformer implementation="org.apache.maven.plugins.shade.resource.AppendingTransformer">
                   <resource>META-INF/kie.conf</resource>
                 </transformer>
-                <transformer
-                    implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                   <mainClass>org.openjdk.jmh.Main</mainClass>
                 </transformer>
               </transformers>
@@ -218,7 +193,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
       </plugin>
     </plugins>
   </build>
-
 
   <profiles>
     <profile>
@@ -240,7 +214,6 @@ THE POSSIBILITY OF SUCH DAMAGE.
         <maven.jdbc.db.port/>
         <maven.jdbc.db.server/>
         <maven.jdbc.url>jdbc:h2:mem:test;MVCC=true</maven.jdbc.url>
-
       </properties>
     </profile>
   </profiles>
@@ -248,11 +221,7 @@ THE POSSIBILITY OF SUCH DAMAGE.
   <properties>
     <!-- Tomcat DBCP 2 pool configuration -->
     <maven.dbcp2.poolSize>16</maven.dbcp2.poolSize>
-
-    <kit.version>${version.org.kie}</kit.version>
-    <metrics.version>3.1.0</metrics.version>
   </properties>
-
 
 </project>
