+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/filtered-resources/datasource.properties
@@ -1,6 +1,6 @@
 # JDBC/Database properties that are set in the maven pom
 #
-#   the below variable names (i.e. "${maven.datasource.classname}) are
+# The below variable names (i.e. "${maven.datasource.classname}) are
 # automatically replaced with their values (defined in the pom.xml)
 # because of the fact that <filtering> is set to true in for the
 # src/main/resources directory in the pom.
