+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/java/com/redhat/bpms/examples/mortgage/Appraisal.java
@@ -1,4 +1,4 @@
-package mortgage;
+package com.redhat.bpms.examples.mortgage;
 
 /**
  * This class was automatically generated by the data modeler tool.
