+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/java/org/jbpm/test/performance/jbpm/constant/ProcessStorage.java
@@ -4,25 +4,20 @@ public enum ProcessStorage {
 
     HumanTask("HumanTask.bpmn2", "org.kie.perf.HumanTask"),
     GroupHumanTask("GroupHumanTask.bpmn2", "org.kie.perf.GroupHumanTask"),
-    ParallelGateway(
-            "ParallelGateway.bpmn2", "org.kie.perf.ParallelGateway"),
-    ParallelGatewayTenTimes("ParallelGatewayTenTimes.bpmn2",
-                            "org.kie.perf.ParallelGatewayTenTimes"),
+    ParallelGateway("ParallelGateway.bpmn2", "org.kie.perf.ParallelGateway"),
+    ParallelGatewayTenTimes("ParallelGatewayTenTimes.bpmn2", "org.kie.perf.ParallelGatewayTenTimes"),
     ParallelGatewayTwoTimes("ParallelGatewayTwoTimes.bpmn2", "org.kie.perf.ParallelGatewayTwoTimes"),
-    RuleTask(
-            "RuleTask.bpmn2", "org.kie.perf.RuleTask"),
+    RuleTask("RuleTask.bpmn2", "org.kie.perf.RuleTask"),
     ScriptTask("ScriptTask.bpmn2", "org.kie.perf.ScriptTask"),
-    IntermediateSignal(
-            "IntermediateSignal.bpmn2", "org.kie.perf.IntermediateSignal"),
+    IntermediateSignal("IntermediateSignal.bpmn2", "org.kie.perf.IntermediateSignal"),
     StartEnd("StartEnd.bpmn2", "org.kie.perf.StartEnd"),
-    IntermediateTimer(
-            "IntermediateTimer.bpmn2", "org.kie.perf.IntermediateTimer"),
+    IntermediateTimer("IntermediateTimer.bpmn2", "org.kie.perf.IntermediateTimer"),
     MortgageApplication("MortgageApplication.bpmn2", "com.redhat.bpms.examples.mortgage.MortgageApplication");
 
     private String path;
     private String processDefinitionId;
 
-    private ProcessStorage(String name, String processDefinitionId) {
+    ProcessStorage(String name, String processDefinitionId) {
         this.path = "processes/" + name;
         this.processDefinitionId = processDefinitionId;
     }
