+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/java/org/jbpm/test/performance/jbpm/constant/RuleStorage.java
@@ -2,12 +2,11 @@ package org.jbpm.test.performance.jbpm.constant;
 
 public enum RuleStorage {
 
-    ValidationUserFact("ValidationUserFact.drl"),
-    ;
+    ValidationUserFact("ValidationUserFact.drl");
 
     private String path;
 
-    private RuleStorage(String name) {
+    RuleStorage(String name) {
         this.path = "rules/" + name;
     }
