+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/java/org/jbpm/test/performance/scenario/load/LHumanTaskProcessWithListeners.java
@@ -29,7 +29,6 @@ import org.openjdk.jmh.annotations.Warmup;
 @Warmup(iterations = 1, time = 1)
 @Measurement(iterations = 1, time = 1)
 @Threads(1)
-
 public class LHumanTaskProcessWithListeners {
 
     // ! Must be overridden using -p from command line
