+++ jbpm-benchmarks/jbpm-performance-tests-jmh/src/main/java/org/jbpm/test/performance/scenario/load/LParallelGatewayTwoTimesProcess.java
@@ -26,7 +26,6 @@ import org.openjdk.jmh.annotations.Warmup;
 @Warmup(iterations = 1, time = 1)
 @Measurement(iterations = 1, time = 1)
 @Threads(1)
-
 public class LParallelGatewayTwoTimesProcess {
 
     // ! Must be overridden using -p from command line
