+++ jbpm-benchmarks/jbpm-performance-tests/src/main/java/org/jbpm/test/performance/scenario/PrepareEngine.java
@@ -9,6 +9,7 @@ import org.jbpm.services.task.impl.factories.TaskFactory;
 import org.kie.api.task.TaskService;
 import org.kie.api.task.model.Task;
 
+
 public class PrepareEngine {
 
     public static List<Long> createNewTasks(boolean start, int count, TaskService taskService, String runtimeManagerIdentifier) {
@@ -16,7 +17,7 @@ public class PrepareEngine {
         str += "peopleAssignments = (with ( new PeopleAssignments() ) { potentialOwners = [new User('" + UserStorage.PerfUser.getUserId()
                 + "')], businessAdministrators = [ new User('Administrator') ], }),";
         str += "names = [ new I18NText( 'en-UK', 'perf-sample-task')] })";
-        List<Long> taskIds = new ArrayList<Long>();
+        List<Long> taskIds = new ArrayList<>();
         while (count > 0) {
             Task task = TaskFactory.evalTask(new StringReader(str));
             long taskId = taskService.addTask(task, null);
